Src/deploy/requirements.txt
flask
bootstrap-flask
Pillow
numpy
torch
ultralytics
camelot-py
opencv-python
pandas
openpyxl
pdf2image
python-magic
pytesseract
gunicorn
python-dotenv
requests
scikit-image
matplotlib
zipfile
